// src/pages/user/lottery/interface.tsx
interface IGoods {
  app: number;
  appPrice0: number;
  appPrice1: number;
  appPrice2: number;
  appUsable: number;
  colorId: number;
  commonId: number;
  giftVoList: any[];
  goodsName: string;

  // goodsCommon: IGoodsCommon;
  goodsFullSpecs: string;
  goodsId: number;
  goodsPrice: number;
  goodsPrice0: number;
  goodsPrice1: number;
  goodsPrice2: number;
  goodsSerial: any;
  goodsSpecString: string;
  goodsSpecs: string;
  goodsStorage: number;
  groupPrice: number;
  imageName: string;
  imageSrc: string;
  isGift: number;
  isGroup: any;
  isSeckill: number;
  limitAmount: number;
  promotionEndTime: string;
  promotionId: number;
  promotionStartTime: string;
  promotionState: number;
  promotionTitle: string;
  promotionType: number;
  promotionTypeText: string;
  specValueIds: string;
  web: number;
  webPrice0: number;
  webPrice1: number;
  webPrice2: number;
  webUsable: number;
  wechat: number;
  wechatPrice0: number;
  wechatPrice1: number;
  wechatPrice2: number;
  wechatUsable: number;
}

interface ISpecialBatch {
  activityId: any;
  allowSubscribe: boolean;
  autoStartEnd: any;
  beginTime: string;
  commonId: any;
  computeRuleUrl: any;
  contentUrl: any;
  createBy: any;
  createDate: any;
  denyCode: any;
  denyDesc: any;
  detailImage1: any;
  detailImage1_1: any;
  detailImage2: any;
  detailImage2_2: any;
  detailImage3: any;
  detailImage3_3: any;
  detailImage4: any;
  detailImage4_4: any;
  drawTime: string;
  endTime: string;
  goodsId: any;
  goodsPrice: any;
  id: any;
  isDel: any;
  isExpired: any;
  isPay: any;
  isWin: any;
  issueQuantity: any;
  mainImage1: any;
  mainImage2: any;
  mainImage3: any;
  modifyBy: any;
  modifyDate: any;
  orderId: any;
  orderSn: any;
  payExpireTime: string;
  payId: any;
  payQuantity: any;
  phone: any;
  remainBeginTime: number;
  remainDrawTime: number;
  remainPayTime: number;
  remainSubscribeTime: number;
  sendQuantity: any;
  state: any;
  subscribeQuantity: any;
  subscribeState: any;
  subscribeTime: any;
  subscriberSn: any;
}

export interface ISpecialListItem {
  goods: IGoods;
  specialtyBatch: ISpecialBatch;
  activityId: number;
  batchId: number;
  commonId: number;
  computeId: any;
  createBy: any;
  createDate: string;
  goodsId: number;
  id: number;
  isDel: number;
  isExpired: any;
  isPay: number;
  isWin: number;
  modifyBy: any;
  modifyDate: any;
  orderId: number;
  orderSn: number;
  payId: number;
  payTime: any;
  storeName: string;
  subscribeState: number;
  subscribeStateStr: string;
  subscribeTime: string;
  subscriberId: number;
  subscriberSn: number;
}